refactoring path decorator to inspect controllers
this change moves away from the previous usage of path and method
decorators. the path decorator continues to be used for decorating
controller classes, and will inspect the controller class to build up
the method paths. the method decorator is disabled until it can be
incorporated into the new format.

changes
* updating readme to reflect new changes
* updating example doc app to use pretty print package
* updating example app to show functionality of changes
* adding function to collect paths from a list of controllers
* adding function to collect controllers from a path
* refactoring get_paths to use new controllers
* refactoring methods_get to look for index
* fixing path join
* adding controllers build function to utils
* adding paths build function to utils